+++ agent/impl/src/main/java/com/walmartlabs/concord/agent/ExecutionManager.java
@@ -12,10 +12,13 @@ import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.Inet6Address;
+import java.net.InetAddress;
+import java.net.NetworkInterface;
+import java.net.SocketException;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.*;
 import java.util.concurrent.TimeUnit;
 import java.util.zip.ZipInputStream;
 
@@ -58,6 +61,8 @@ public class ExecutionManager {
         // e.g. left after the execution was suspended
         logManager.delete(instanceId);
         logManager.touch(instanceId);
+
+        logManager.log(instanceId, "Agent IPs: %s", String.join(", ", getLocalIPs()));
 
         Path tmpDir = extract(payload);
 
@@ -141,5 +146,30 @@ public class ExecutionManager {
         } catch (IOException e) {
             throw new ExecutionException("Error while unpacking a payload", e);
         }
+    }
+
+    private static Set<String> getLocalIPs() {
+        Set<String> result = new HashSet<>();
+
+        try {
+            Enumeration<NetworkInterface> ifs = NetworkInterface.getNetworkInterfaces();
+            while (ifs.hasMoreElements()) {
+                NetworkInterface i = ifs.nextElement();
+                Enumeration<InetAddress> addrs = i.getInetAddresses();
+                while (addrs.hasMoreElements()) {
+                    InetAddress a = addrs.nextElement();
+                    if (a instanceof Inet6Address) {
+                        // skip IPv6 addresses
+                        continue;
+                    }
+                    result.add(a.getHostAddress());
+                }
+            }
+        } catch (SocketException e) {
+            log.warn("getLocalIPs -> can't determine the local IP addresses: {}", e.getMessage());
+            return Collections.singleton("n/a");
+        }
+
+        return result;
     }
 }